Template fixes & Added HellDivers icons
Added transparent HellDivers icons (Left & Right).
Corrected line number in .ini template files.
Fixed "Information" tag, as when replacing "NAMEHERE" with the icon name ("HellDivers Left" for example") it would end up saying "HellDivers Left icon that is shown on the right side"; So, yeah, it was a bit redundant so I fixed it.

If you're reading this, I hope you have a great day!